requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
pypdf
chromadb
faiss-cpu
google.generativeai
langchain_google_genai

